fix: implement project delete
